requirements.txt
# Data ingestion
yfinance==0.2.40
requests==2.31.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Data processing
pandas==2.2.2
numpy==1.26.4
ta

# Scheduling
APScheduler==3.10.4

# Environment
python-dotenv==1.0.1

# Forecasting (Phase 4 - install when ready)
# prophet==1.1.5
# statsmodels==0.14.2

# Experiment tracking (Phase 4)
# mlflow==2.13.0

# Dashboard (Phase 5)
# streamlit==1.35.0
# plotly==5.22.0

# Alerts (Phase 5)
# slack-sdk==3.27.2
